requirements.txt
absl-py==0.7.0
aniso8601==3.0.2
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.2
cycler==0.10.0
decorator==4.3.0
Django==1.11.22
django-filter==2.0.0
factory-boy==2.11.1
Faker==1.0.4
gast==0.2.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
grpcio==1.18.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.7
inotify==0.2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
mock==2.0.0
more-itertools==5.0.0
numpy==1.16.0
olefile==0.46
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.8.1
promise==2.2.1
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.6
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
py==1.7.0
Pygments==2.3.1
pyparsing==2.3.1
pyshp==2.0.1
pytest==4.1.1
pytest-django==3.4.5
python-dateutil==2.7.5
python-redis-lock==3.2.0
pytz==2018.9
redis==3.0.1
requests==2.20.1
rq==0.13.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
Werkzeug==0.14.1